heuristics are now used when we find equal values while comparing at getMove().

+++ src/de/tudresden/inf/ggp/basicplayer/IHeuristic.java
@@ -3,5 +3,5 @@ package de.tudresden.inf.ggp.basicplayer;
 import org.eclipse.palamedes.gdl.core.model.IGameNode;
 
 public interface IHeuristic {
-	public int calculateHeuristic(IGameNode node);
+	public double calculateHeuristic(IGameNode node);
 }
