+++ src/de/tudresden/inf/ggp/basicplayer/InverseMobilityHeuristic.java
@@ -7,7 +7,7 @@ public class InverseMobilityHeuristic implements IHeuristic {
 	private int roleIndex;
 	
 	@Override
-	public int calculateHeuristic(IGameNode node) {
+	public double calculateHeuristic(IGameNode node) {
 		int myMoves = node.getState().getLegalMoves()[roleIndex].length;
 		int movesOfTheOthers = 0;
 		for(int i=0; i<node.getState().getLegalMoves().length; i++) {
